Ensure client names match up with actual endpoints
